Change the Ball class size variable to float type instead of int type

+++ ColorBalls/Ball.cs
@@ -7,17 +7,17 @@ namespace ColorBalls
     class Ball
     {
         private Colour colour;
-        private int size;
+        private float size;
         private int throwCount;
 
-        public Ball(byte rVal, byte gVal, byte bVal, int size)
+        public Ball(byte rVal, byte gVal, byte bVal, float size)
         {
             this.colour = new Colour(rVal, gVal, bVal);
             this.size = size;
             this.throwCount = 0;
         }
 
-        public int Pop() => this.size = 0;
+        public float Pop() => this.size = 0;
 
         public int Throw()
         {
